feat: app refactoring + add Main Interface Component + set bet input field
